requirements.txt
fastapi==0.68.2
uvicorn==0.15.0
sqlalchemy==1.4.26
alembic==1.7.4
python-dotenv==0.19.1
pymysql==1.0.2
passlib

